+++ lavis/models/blip2_models/blip2_t5.py
@@ -84,7 +84,7 @@ class Blip2T5(Blip2Base):
 
         for name, param in self.t5_model.named_parameters():
             param.requires_grad = False
-            param.data = param.data.bfloat16()
+            param.data = param.data.float()
 
         self.t5_proj = nn.Linear(
             self.Qformer.config.hidden_size, self.t5_model.config.hidden_size
